+++ rpgJavaInterpreter-core/src/main/kotlin/com/smeup/rpgparser/parsing/ast/PredefinedIndicatorExpr.kt
@@ -5,7 +5,7 @@ import com.strumenta.kolasu.model.Position
 // *IN00..*IN99
 data class PredefinedIndicatorExpr(val index: Int, override val position: Position? = null) :
     AssignableExpression(position) {
-    override fun size(): Long {
+    override fun size(): Int {
         TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
     }
 }
@@ -13,7 +13,7 @@ data class PredefinedIndicatorExpr(val index: Int, override val position: Positi
 // *IN
 data class PredefinedGlobalIndicatorExpr(override val position: Position? = null) :
         AssignableExpression(position) {
-    override fun size(): Long {
+    override fun size(): Int {
         TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
     }
 }
